requirements.txt
phidata
groq
langchain
yfinance
dotenv
duckduckgo-search
fastapi
uvicorn 
packaging
openai